requirements.txt
kafka-python-ng
yfinance
pandas
requests